+++ CHANGELOG.md
@@ -6,4 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+0.0.5
+
+- debouncing of concurrent invocations of puku for the same package
+- call puku fmt with go filename rather than dirname